Refactor include filter module into filter package
